MENT-409 Non-blocking DDL implementation
Non-blocking DDL is implemented by two-phase TOI method. The first
TOI phase grabs necessary MDL locks and releases commit order critical
section before proceeding to the actual DDL execution. The DDL
execution proceeds in BF mode, so it cannot be interrupted.
The second phase starts when the DDL execution finishes. A new
TOI is grabbed and held until binlogging and MDL release is over.

On applier side a new NBO worker thread is started for handling the
DDL. The applier thread waits until the worker thread grabs MDL locks
and then returns and releases TOI critical section. After this the
worker thread processes the DDL in the same way as regular client
session.

A wsrep_RSU_mode = NBO has been added, which needs to be set by the
user before running the DDL in order to make the DDL run in non-blocking
mode.

Initially the following DDLs are allowed in NBO mode:

- SQLCOM_CREATE_INDEX with locking modes ALTER_TABLE_LOCK_SHARED and
  ALTER_TABLE_LOCK_EXCLUSIVE.
- SQLCOM_DROP_INDEX.
- SQLCOM_OPTIMIZE.

Co-authored-by: Leandro Pacheco <leandro.pacheco@galeracluster.com>