MENT-408: Galera Black Box for troubleshooting
This is the first version of the Black Box (BB) feature that extends
the logging mechanism of MySQL and MariaDB database servers. It saves
all log messages regardless of the log level setting for a short
time. These log messages are stored in the BB until it becomes
full. After that new log messages overwrite the oldest ones in BB.

Feature developed by Pekka Lampio (pekka.lampio@galeracluster.com)